feat(middleware): inclusão do middleware 'authentication'
